COM_HashForString: Fix undefined signed int wrapping

+++ code/game/q_shared.h
@@ -994,7 +994,7 @@ void PerpendicularVector( vec3_t dst, const vec3_t src );
 
 float Com_Clamp( float min, float max, float value );
 
-int		COM_HashForString(const char *str, int size);
+unsigned	COM_HashForString(const char *str, size_t size);
 char	*COM_SkipPath( char *pathname );
 void	COM_StripExtension( const char *in, char *out, size_t destsize );
 void	COM_DefaultExtension( char *path, int maxSize, const char *extension );
